requirements.txt
aniso8601==9.0.1
appnope==0.1.2
APScheduler==3.6.3
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bidict==0.21.2
bleach==3.3.0
cachetools==4.2.2
camelot-py==0.10.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
cryptography==3.4.8
cycler==0.10.0
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
distro==1.6.0
Django==3.1.7
entrypoints==0.3
et-xmlfile==1.1.0
euclid3==0.1
Flask==1.1.2
Flask-RESTful==0.3.9
Flask-SocketIO==5.0.1
future==0.18.2
idna==3.2
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.5.2
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
loguru==0.5.3
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.3.4
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
multiprocess==0.70.12.2
multitasking==0.0.9
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.0
numpy==1.20.1
oauthlib==3.1.0
opencv-python==4.5.3.56
openpyxl==3.0.7
packaging==20.9
pandas==1.2.3
pandas-datareader==0.9.0
pandocfilters==1.4.3
parso==0.8.2
pathos==0.2.8
pdfminer.six==20200517
pdfplumber==0.5.28
peewee==3.14.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pox==0.3.0
ppft==1.6.6.4
prometheus-client==0.11.0
prompt-toolkit==3.0.18
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
pycryptodome==3.10.1
PyExecJS==1.5.1
pygame==2.0.1
pygcode==0.1.0
Pygments==2.9.0
pylint==2.6.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-engineio==4.0.0
python-pptx==0.6.19
python-socketio==5.0.4
python-telegram-bot==13.7
pytz==2021.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.1
tabula-py==2.3.0
tabulate==0.8.9
terminado==0.10.1
testpath==0.5.0
toml==0.10.1
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
translator==0.0.9
translators==4.9.5
tweepy==3.9.0
tzlocal==3.0
urllib3==1.26.6
Wand==0.6.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
XlsxWriter==3.0.1
yahoofinancials==1.6
yfinance==0.1.55